requirements.txt
beautifulsoup4==4.11.2
favicon==0.7.0
joblib==1.2.0
numpy==1.24.2
pandas==1.5.3
pycountry==22.3.5
python_whois==0.8.0
Requests==2.31.0
scikit_learn==1.2.2
streamlit==1.22.0
tldextract==3.4.0
